+++ mobile/android/android-components/components/lib/crash-sentry/src/test/java/mozilla/components/lib/crash/sentry/SentryServiceTest.kt
@@ -22,6 +22,7 @@ import mozilla.components.concept.base.crash.Breadcrumb as MozillaBreadcrumb
 
 @RunWith(AndroidJUnit4::class)
 class SentryServiceTest {
+    class TestException : Exception()
 
     @Before
     fun setup() {